% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeExcept.R
\name{removeExcept}
\alias{removeExcept}
\title{Remove all items objects from global environment except those specified by user.}
\usage{
removeExcept(x)
}
\arguments{
\item{x}{character}
}
\value{

}
\description{
Remove all items objects from global environment except those specified by user.
}
\examples{
firstObject <- c(a, b, c)
secondObject <- c(x, y, z)
removeExcept("secondObject")
}
